refactor: move dbsync mock to dbsync folder
