+++ src/wazuh_modules/sca/sca_impl/tests/mocks/sca_event_handler_mock.hpp
@@ -4,7 +4,7 @@
 
 #include <sca_event_handler.hpp>
 
-#include "mockdbsync.hpp"
+#include <mock_dbsync.hpp>
 
 namespace sca_event_handler
 {
