+++ src/wazuh_modules/sca/sca_impl/tests/sca_test.cpp
@@ -4,7 +4,7 @@
 
 #include <dbsync.hpp>
 
-#include "mocks/mockdbsync.hpp"
+#include <mock_dbsync.hpp>
 
 #include <memory>
 #include <string>
